+++ CHANGELOG.md
@@ -3,12 +3,17 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
+## [0.2.4] - 2025-11-07
+
+### Added
+
+- Added cutting treatment #MELA2-104
 
 ## [0.2.3] - 2025-11-07
 
 ### Added
 
-- Added cutting treatment #MELA2-104
+- Added regeneration treatment and planting event
 
 ## [0.2.2] - 2025-11-07